refactor: (#87) Swagger 인터페이스명 수정
